Remove unnecessary `as` casts

+++ bfs-3x3-2-color-ufb/src/cube.rs
@@ -359,8 +359,8 @@ impl Cube {
 impl From<CoordCube<'_>> for Cube {
     fn from(value: CoordCube<'_>) -> Self {
         let mut cube = Cube::new();
-        cube.set_edges_coord(value.edges as u32);
-        cube.set_corners_coord(value.corners as u32);
+        cube.set_edges_coord(value.edges);
+        cube.set_corners_coord(value.corners);
         cube
     }
 }
